+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 # unreleased
 * Add opt-in setting for future features
 * Fix issue where modal backdrop did not freeze
+* close modal/drawer when backdrop is clicked
 
 # 0.1.0 - Pre-release
 * Add EDHRec Suggestions Button